requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
pgvector>=0.2.3
redis>=5.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=3.2.2,<5
python-multipart==0.0.6
httpx>=0.25.2
celery>=5.3.4
langchain>=0.1.0
langgraph>=0.0.53
langchain-openai>=0.0.8
openai>=1.6.1
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
gspread>=5.12.4
pandas>=2.1.4
numpy>=1.26.0
python-docx>=1.1.0
PyPDF2>=3.0.1
python-pptx>=0.6.21
xlrd>=2.0.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1
prometheus-client>=0.19.0
structlog>=23.2.0
